requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.6
billiard==3.6.4.0
black==22.1.0
blue==0.9.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorlog==6.6.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-sslserver==0.22
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.7.0
filelock==3.7.1
flake8==4.0.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
pathspec==0.9.0
pika==1.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.15.0
pyflakes==2.4.0
PyJWT==1.7.1
pylint==2.14.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-ldap==3.4.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
types-Markdown==3.3.28
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.31
types-urllib3==1.26.15
typing_extensions==4.2.0
ujson==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
